rename 이름 수정작업 & chore review.ts에 reviewDetail.ts 병합
